# apps/messaging-api/openapi-definition.yml
openapi: 3.0.3
info:
  title: OGCIO MessagingIE API
  description: API for OGCIO MessagingIE Service
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /api/v1/messages/:
    get:
      operationId: ListMessages
      tags: &a1
        - Messages
      description: Returns all the messages for the requested organisation or the
        requested recipient
      parameters:
        - schema:
            type: string
            enum:
              - delivered
          in: query
          name: status
          required: false
        - schema:
            type: string
            enum:
              - "true"
              - "false"
              - "0"
              - "1"
          in: query
          name: isSeen
          required: false
        - schema:
            type: string
          in: query
          name: search
          required: false
        - schema:
            type: string
            enum:
              - "true"
              - "false"
              - "0"
              - "1"
          in: query
          name: deleted
          required: false
        - schema:
            type: string
          in: query
          name: recipientUserId
          required: false
          description: Either recipientUserId and organisationId are mandatory
        - schema:
            type: string
          in: query
          name: organisationId
          required: false
          description: Either recipientUserId and organisationId are mandatory
        - schema:
            type: string
            pattern: ^[0-9][0-9]*|undefined$
            default: "0"
          in: query
          name: offset
          required: false
          description: Indicates where to start fetching data or how many records to skip,
            defining the initial position within the list
        - schema:
            type: string
            default: "20"
            pattern: ^([1-9]|100)|undefined$
          in: query
          name: limit
          required: false
          description: Indicates the maximum number (100) of items that will be returned
            in a single request
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - subject
                        - createdAt
                        - threadName
                        - organisationId
                        - recipientUserId
                        - attachmentsCount
                      properties:
                        id:
                          type: string
                          description: Unique Id of the message
                        subject:
                          type: string
                          description: Subject
                        createdAt:
                          type: string
                          description: Creation date time
                        threadName:
                          anyOf:
                            - type: string
                              description: Thread Name used to group messages
                            - type: "null"
                        organisationId:
                          type: string
                          description: Organisation sender id
                        recipientUserId:
                          type: string
                          description: Unique id of the recipient
                        attachmentsCount:
                          type: integer
                          description: Number of attachments
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        required:
                          - self
                          - first
                          - last
                          - pages
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the request itself
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the next page of results in a paginated response.
                                  If there are no more results, this field may
                                  be omitted
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the previous page of results in a paginated
                                  response. If there are no more results, this
                                  field may be omitted
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          pages:
                            type: object
                            description: It may contain a list of other useful URLs, e.g. one entry for
                              page:'page 1', 'page 2'
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        description: Object containing the links to the related endpoints
                      totalCount:
                        type: number
                        description: Number representing the total number of available items
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
    post:
      operationId: CreateMessage
      tags: *a1
      description: Creates a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - preferredTransports
                - recipientUserId
                - security
                - scheduleAt
                - message
              properties:
                preferredTransports:
                  type: array
                  items:
                    type: string
                    enum:
                      - email
                      - lifeEvent
                    description: All the available provider types
                  description: The list of the preferred transports to use. If the selected
                    transports are not available for the recipient, others will
                    be used
                recipientUserId:
                  type: string
                  description: Unique user id of the recipient
                security:
                  type: string
                  enum:
                    - confidential
                    - public
                  description: Confidentiality level of the message
                scheduleAt:
                  type: string
                  format: date-time
                  description: Date and time of when schedule the message
                message:
                  type: object
                  required:
                    - subject
                    - plainText
                    - language
                  properties:
                    threadName:
                      type: string
                      description: Thread Name used to group messages
                      minLength: 1
                    subject:
                      type: string
                      description: Subject. This is the only part that will be seen outside of the
                        messaging platform is security is 'confidential'
                      minLength: 1
                    excerpt:
                      type: string
                      description: Brief description of the message
                    plainText:
                      type: string
                      description: Plain text version of the message
                      minLength: 1
                    richText:
                      type: string
                      description: Rich text (html) version of the message
                    language:
                      type: string
                      enum:
                        - en
                        - ga
                      description: Language used to send the message
                    externalId:
                      type: string
                      description: External id to link the message to an external system
                attachments:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - id
                    properties:
                      id:
                        type: string
                        format: uuid
                        description: Unique Id of the resource
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        required:
                          - self
                          - first
                          - last
                          - pages
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the request itself
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the next page of results in a paginated response.
                                  If there are no more results, this field may
                                  be omitted
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the previous page of results in a paginated
                                  response. If there are no more results, this
                                  field may be omitted
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          pages:
                            type: object
                            description: It may contain a list of other useful URLs, e.g. one entry for
                              page:'page 1', 'page 2'
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        description: Object containing the links to the related endpoints
                      totalCount:
                        type: number
                        description: Number representing the total number of available items
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
    delete:
      tags: *a1
      description: Deletes the message with the given message id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: List of message ids to delete
                  minItems: 1
                  maxItems: 100
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - ids
                    properties:
                      ids:
                        type: array
                        items:
                          type: string
                          format: uuid
                        description: List of message ids to delete
                        minItems: 1
                        maxItems: 100
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
...
security:
  - bearerAuth: []